+++ Core/GenieError.cs
@@ -2,7 +2,12 @@
 
 namespace GenieClient
 {
-    public static class GenieError
+    /// <summary>
+    /// Central error handling for the Genie client.
+    /// Base class provides general error events. Plugin-specific error handling
+    /// is extended in the Windows project via partial class.
+    /// </summary>
+    public static partial class GenieError
     {
         public static void Error(string section, string message, string description = null)
         {
@@ -12,20 +17,5 @@ namespace GenieClient
         public static event EventGenieErrorEventHandler EventGenieError;
 
         public delegate void EventGenieErrorEventHandler(string section, string message, string description);
-
-        public static event EventGenieLegacyPluginErrorEventHandler EventGenieLegacyPluginError;
-        public static event EventGeniePluginErrorEventHandler EventGeniePluginError;
-
-        public delegate void EventGenieLegacyPluginErrorEventHandler(GeniePlugin.Interfaces.IPlugin plugin, string section, Exception ex);
-        public delegate void EventGeniePluginErrorEventHandler(GeniePlugin.Plugins.IPlugin plugin, string section, Exception ex);
-
-        public static void GeniePluginError(GeniePlugin.Interfaces.IPlugin plugin, string section, Exception ex)
-        {
-            EventGenieLegacyPluginError?.Invoke(plugin, section, ex);
-        }
-        public static void GeniePluginError(GeniePlugin.Plugins.IPlugin plugin, string section, Exception ex)
-        {
-            EventGeniePluginError?.Invoke(plugin, section, ex);
-        }
     }
 }
