--- Genie4.sln
@@ -1,49 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.32014.148
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{F184B08F-C81C-45F6-A57F-5ABD9991F28F}") = "Plugins", "Plugin\Plugins.vbproj", "{B0882700-5EAF-4712-ADE5-69DAFE473A01}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Genie4", "Genie4.csproj", "{DA91259E-E378-443E-A0DE-11E8E160842C}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|Mixed Platforms = Debug|Mixed Platforms
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|Mixed Platforms = Release|Mixed Platforms
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B0882700-5EAF-4712-ADE5-69DAFE473A01}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B0882700-5EAF-4712-ADE5-69DAFE473A01}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B0882700-5EAF-4712-ADE5-69DAFE473A01}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{B0882700-5EAF-4712-ADE5-69DAFE473A01}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{B0882700-5EAF-4712-ADE5-69DAFE473A01}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{B0882700-5EAF-4712-ADE5-69DAFE473A01}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B0882700-5EAF-4712-ADE5-69DAFE473A01}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B0882700-5EAF-4712-ADE5-69DAFE473A01}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{B0882700-5EAF-4712-ADE5-69DAFE473A01}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{B0882700-5EAF-4712-ADE5-69DAFE473A01}.Release|x86.ActiveCfg = Release|Any CPU
-		{DA91259E-E378-443E-A0DE-11E8E160842C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DA91259E-E378-443E-A0DE-11E8E160842C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DA91259E-E378-443E-A0DE-11E8E160842C}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{DA91259E-E378-443E-A0DE-11E8E160842C}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{DA91259E-E378-443E-A0DE-11E8E160842C}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{DA91259E-E378-443E-A0DE-11E8E160842C}.Debug|x86.Build.0 = Debug|Any CPU
-		{DA91259E-E378-443E-A0DE-11E8E160842C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DA91259E-E378-443E-A0DE-11E8E160842C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DA91259E-E378-443E-A0DE-11E8E160842C}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{DA91259E-E378-443E-A0DE-11E8E160842C}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{DA91259E-E378-443E-A0DE-11E8E160842C}.Release|x86.ActiveCfg = Release|Any CPU
-		{DA91259E-E378-443E-A0DE-11E8E160842C}.Release|x86.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {0CD56CDA-CF3F-4592-99EE-F961EBDE2D53}
-	EndGlobalSection
-EndGlobal
