+++ build.ps1
@@ -1,13 +1,13 @@
-# Quick Build Script for Genie4
+# Quick Build Script for Genie5
 # Builds and copies to Run folder (no prompts)
 
 $BuildConfig = "Release"
 $TargetFramework = "net10.0-windows"
-$SourceDir = "bin\$BuildConfig\$TargetFramework"
+$SourceDir = "src\Genie.Windows\bin\$BuildConfig\$TargetFramework"
 $RunDir = "bin\Run"
 
-Write-Host "Building Genie4..." -ForegroundColor Cyan
-dotnet build Genie4.sln --configuration $BuildConfig
+Write-Host "Building Genie (Core + Windows)..." -ForegroundColor Cyan
+dotnet build Genie5.sln --configuration $BuildConfig
 
 if ($LASTEXITCODE -ne 0) {
     Write-Host "Build failed!" -ForegroundColor Red
